Move to `loc.js`
