+++ src/language-js/loc.js
@@ -1,5 +1,13 @@
 "use strict";
 
+/**
+ * @typedef {import("./types/estree").Node} Node
+ */
+
+/**
+ * @param {Node} node
+ * @returns {number}
+ */
 function locStart(node, opts) {
   const { ignoreDecorators } = opts || {};
 
@@ -16,11 +24,20 @@ function locStart(node, opts) {
   return node.range ? node.range[0] : node.start;
 }
 
+/**
+ * @param {Node} node
+ * @returns {number}
+ */
 function locEnd(node) {
   const end = node.range ? node.range[1] : node.end;
   return node.typeAnnotation ? Math.max(end, locEnd(node.typeAnnotation)) : end;
 }
 
+/**
+ * @param {Node} startNode
+ * @param {Node | number} endNodeOrLength
+ * @returns {number[]}
+ */
 function composeLoc(startNode, endNodeOrLength = startNode) {
   const start = locStart(startNode);
   const end =
@@ -31,8 +48,37 @@ function composeLoc(startNode, endNodeOrLength = startNode) {
   return [start, end];
 }
 
+/**
+ * @param {Node} nodeA
+ * @param {Node} nodeB
+ * @returns {boolean}
+ */
+function hasSameLocStart(nodeA, nodeB) {
+  return locStart(nodeA) === locStart(nodeB);
+}
+
+/**
+ * @param {Node} nodeA
+ * @param {Node} nodeB
+ * @returns {boolean}
+ */
+function hasSameLocEnd(nodeA, nodeB) {
+  return locEnd(nodeA) === locEnd(nodeB);
+}
+
+/**
+ * @param {Node} nodeA
+ * @param {Node} nodeB
+ * @returns {boolean}
+ */
+function hasSameLoc(nodeA, nodeB) {
+  return hasSameLocStart(nodeA, nodeB) && hasSameLocEnd(nodeA, nodeB);
+}
+
 module.exports = {
   locStart,
   locEnd,
   composeLoc,
+  hasSameLocStart,
+  hasSameLoc,
 };
